Speedup cancellation at the top level
Multiplying a rational function by the denominator is slow. This avoids
having the denominator in the first place resulting in a considerable
speedup. This also tests the idea of recursing with two different spins.
Perhaps this is why the c_1, c_2 and c_3 residues are not quoted as
explicit functions of spin.

Signed-off-by: Connor Behan <connor.behan@gmail.com>