Fix huge gap in my understanding
My idea to pass two different spins in the recurrence relation was
WRONG. Divergences that occur with only one spin are SUPPOSED to occur.
It is a consequence of the fact that even-dimensional blocks have double
poles. It just takes care to notice that even though huge coefficients
do not occur in pairs, they will still cancel after looking at all
recursion depths.

Signed-off-by: Connor Behan <connor.behan@gmail.com>