Add more paranoia
First, the gap being imposed during a bisect needs to be stored at high
precision. Second, the unit operator had some garbage in it. I tried to
take too much of a shortcut while calculating it, so this just does the
straightforward thing.

This program is now ready for simple bootstrap applications. The table
is still a bit SLOW but as far as I can tell, all tests yield correct
results.

Signed-off-by: Connor Behan <connor.behan@gmail.com>