Use more native code for arbitrary precision
It is an absolute pleasure to be using a library that is so actively
developed at this time. When I started PyCFTBoot, csympy did not even
include the mpfr type. Now it does and after another short wait, we will
probably be able to use it for the bilinear basis too.

Signed-off-by: Connor Behan <connor.behan@gmail.com>