Make the imposed gaps persistent
Specifying a spin and dimension for shifting upon every XML write is
dumb. Gaps should be set before XML files are written and the SDP type
should remember which gaps have been set until the user resets them to
unitarity. This makes it easy to impose a gap and then find bounds on 2D
OPE coefficients.

Signed-off-by: Connor Behan <connor.behan@gmail.com>