Relate derivatives with coefficients from the Casimirs
Now that one of our co-ordinates runs along the diagonal, the ODE
satisfied by the blocks here leads to much faster evaluation of
derivatives. Off-diagonal derivatives are not implemented yet but this
code is already enough to compare the new and old methods to see that
the same results are produced more quickly.

Signed-off-by: Connor Behan <connor.behan@gmail.com>