Implement off-diagonal Casimir relation
I like this idea. To generate conformal block derivatives quickly, it is
best to seed the Casimir relation with a small number of derivatives
produced via the explicit sum over poles. However, it costs us nothing
to make this seed function capable of generating any number of
derivatives should we need it to. We therefore have redundant generation
of conformal blocks with a fast method and a slow method in case
something goes wrong.

Signed-off-by: Connor Behan <connor.behan@gmail.com>