Fix likely sign error
This redundancy is already paying off. With the old code, the 2D
bootstrap only worked when the slower method was used. This could easily
be the result of a normally negligible error in the double pole code
having its effect amplified by the recurrence relation. After changing a
few parts of the double pole code, this works equally well with both
methods.

Signed-off-by: Connor Behan <connor.behan@gmail.com>