Fix diagonal Casimir relation
While the amount of time that the second Casimir relation can save is
roughly the square of the amount of time that the first Casimir relation
can save, it is important to get both right. A number of things were
wrong before because none of the coefficient polynomials have a degree
above 7. Rather than being differentiated, the change they receive with
each iteration comes from the binomial coefficient they appear beside.

Signed-off-by: Connor Behan <connor.behan@gmail.com>