First stab at global symmetries
To make this work, a trivial change was needed to better handle unit
operators with 0 components. A more drastic change is the way
PolynomialVectors are labelled. Instead of just a spin, they have a
[spin, irrep] pair where irrep is just a label that the user chooses.
This must now be passed whenever assumptions are made on the spectrum.
However, a single integer is assumed to be a spin with irrep = 0
(singlet). Because the position of a particular PolynomialVector in an
SDP table is no longer predictable, several functions just bite the
bullet and search for the one that is needed. When using global
symmetries, not only are there more PolynomialVectors (one channel with
even spins, another channel with odd spins, etc) but there are also more
components in each PolynomialVector. The derivatives of one type of
convolved conformal block are concatenated with the derivatives of
another type.

Signed-off-by: Connor Behan <connor.behan@gmail.com>