Support the extremal functional method
With the last possible functional one finds before a spectrum can no
longer be ruled out, it makes a lot of sense to find the zeros of it.
This allows the user to save the functional found by SDPB and discover
what polynomial it produces when acting on all the conformal block
derivatives. Numpy is used for root finding because it is fast. It is
limited to double precision, but this is much smaller than the error
bars that dimensions have at the end of the day.

Signed-off-by: Connor Behan <connor.behan@gmail.com>