Add routine for extremal OPE coefficients
When an extremal function is found, the number of conformal block
vectors it annihilates will generically be smaller than their dimension.
It therefore makes sense to truncate them. Here we adopt a full
truncation forcing the number of components to be equal to the number of
zeros. In this case, OPE coefficients are found by a simple matrix
inversion. It might be a good idea to use the original strategy of
having one extra degree of freedom which is handled via linear or
semidefinite programming.

Signed-off-by: Connor Behan <connor.behan@gmail.com>