Recurse more efficiently
It makes little sense to calculate meromorphic conformal blocks in a
tree style and once for each spin. This causes the same block to be
computed many times, a problem that should be familiar from Fibonacci
numbers. We should instead compute meromorphic blocks for all spins at
the same time and build them from the bottom up. This gives us a runtime
polynomial in the number of poles instead of the partition function
growth we had earlier.

Signed-off-by: Connor Behan <connor.behan@gmail.com>