Initial support for multiple correlators
This allows a PolynomialVectorMatrix node to actually be a matrix.
Therefore our vectorial sum rule arguments no longer consist of a
coefficient and a 0, 1 for symmetric, antisymmetric blocks. They consist
of a matrix of tuples where each tuple has a coefficient, an index for
which block to use in the passed list and indices for which scaling
dimensions the convolutions are supposed to know about. If things are
passed using the old format, the code lets that slide. However, for some
OPE coefficient tricks that are too hard in the multi-correlator case,
the user is rudely told that the matrix needs to be 1x1.

Signed-off-by: Connor Behan <connor.behan@gmail.com>