Revisit even dimension logic
Ever since the move to multiple correlators, c_2(k) has been unreliable
in even dimension. This investigates how things cancel explicitly. A
trickier problem was figuring out how to handle double poles in 4D. An
infinite residue will not be cancelled by a coincident pair of poles one
level down anymore, but two levels down. This number is expected to
increase with the spacetime dimension. To account for this, we are no
longer allowed to ignore residues that vanish. We must transform them
into a finite part and a zero part. If a coincident pair of poles kills
the zero part at one level, we will be able to probe the next level and
find the next pair of poles that really interests us.

Signed-off-by: Connor Behan <connor.behan@gmail.com>