Choose correctness over speed with even dimensions
Trying to second-guess the recursion relations becomes worse and worse
in 6D and 8D. This keeps the tiny fractional part of the dimension
around as a free symbol and forces it to cancel by keeping all poles /
residues and putting everything over a common denominator. If I have to
change this code one more time, it might be worth removing completely.

Signed-off-by: Connor Behan <connor.behan@gmail.com>