Allow approximating large poles with small poles
For poles with a small residue, a Laurent term can be approximated by a
sum of those for poles with a large residue. This looks at the largest
and smallest dimensions in order to choose optimal coefficients. A
variable "cutoff" is introduced for setting the difference between small
and large. It seems that strange Python conventions require us to call
"import bootstrap" instead of "from bootstrap import *" in order to set
this.

Signed-off-by: Connor Behan <connor.behan@gmail.com>