Enable mass XML generation
We only need orthogonal polynomials with respect to a different measure
when we change the gap. It is a waste to compute them all over again
when we make a new XML file. This calculates the bilinear basis when an
SDP is first made and stores it for future use. It is also sure to
update just the part that will change when a new gap is set. Finally,
SDP types can be chained together to make a new one inherit the bilinear
basis of the old one for faster mixed correlator setups.

Signed-off-by: Connor Behan <connor.behan@gmail.com>