More tutorial errors
Thanks to Junchen Rong for noticing problems in the mixed correlator
example. Apart from an obvious sign error and an overly optimistic
cutoff, I must remember not to mix conformal blocks within a matrix.
Even if the coefficient is zero, some parts of the code infer general
properties from the (0, 0) entry for convenience. The paper will be
updated once this and the SCFT example are sorted out.

Signed-off-by: Connor Behan <connor.behan@gmail.com>