Starting to add the Frobenius method
There are two ways to get the first four derivatives that we plug into
the Casimir ODE. One is to use the Zamolodchikov style recursion
relations, the other is to build a power series that approximately
solves the ODE to some high order. This is an untested addition to
support the second method which makes it easy to work in even d.

Signed-off-by: Connor Behan <connor.behan@gmail.com>