Frobenius fixes
This fixes a few coefficient errors and attempts to cut down the number
of poles. The Frobenius method often generates twice as many poles as
the recurrence relation which means that many of them are unnecessary.
An unnecessary pole is a pole of a conformal block which is also a root
of its numerator. Testing for these and applying a simple factoring
procedure allows SDPB to be sped up. It also removes possible triple
poles which break our integration code and possible poles equal to zero
which break our calculation of the SDP unit. Even after all this, the
Frobenius method appears to have trouble in odd dimensions so we only
call it is clearly the proper choice.

Signed-off-by: Connor Behan <connor.behan@gmail.com>