More reorganization
This concerns the code for co-ordinate changing via the chain rule. It
moves it out of the block files and puts it into the common file. Also
for symmetry, we now have two ways to do the single variable case and
two ways to do the double variable case. One is a hack using Python code
which is currently the fastest. The other is a native Symengine version
and it is perhaps surprising that this is slower. If future work on
Symengine ever reverses this situation, changing which one we use will
be easy.

Signed-off-by: Connor Behan <connor.behan@gmail.com>