Add poles that were missing all this time
Conformal blocks up to spin l_max need to have poles from series 1, 2
and 3. This was fine at the first level, but then some of the residues
above l_max needed to have spins going up even further. Fortunately, the
published islands do not change by even one pixel when I make this
change. But for higher precision studies, this needs to be correct.
Thanks go to Rohit Kalloor for pointing this out.

Signed-off-by: Connor Behan <connor.behan@gmail.com>