Detect bisections that are too long
For quick and dirty bisections, it is nice to only ever have one XML
file. Reuse of the name becomes a problem if checkpoints are generated.
Obviously bisection will break if a checkpoint is loaded in something
that is supposed to be a fresh run of SDPB. This simply times the
calculation and starts using random names if we are found to not be in
the nice case.

Signed-off-by: Connor Behan <connor.behan@gmail.com>