Return a friendlier normalization
My official suggestion is still to divide one OPE coefficient by
another. However, it seems fairly simple to modify the function so that
the true OPE coefficient bound is returned. At least as far as our
normalization of conformal blocks is concerned. There is a common factor
of 4 that people might also like to insert.

Signed-off-by: Connor Behan <connor.behan@gmail.com>