Improve extremal coefficient finding
For all types of operators whose sum rules involve the identity, their
OPE coefficients can be found from a linear system. The old function
played it safe and did not attempt to find this linear system when a
quadratic one (a computational nightmare) was also present. It was also
inadequate in how it made the system square. Extra rows must be
eliminated based on derivative order and extra columns must be
eliminated based on operator dimension. Just going by the order
initially passed to the SDP is bound to lead to problems.

Signed-off-by: Connor Behan <connor.behan@gmail.com>