Change API to find more coefficients
If we make SDP have a better memory of what the external operator
dimensions are, we can exploit OPE coefficient symmetry in order to
solve linear systems given by conformal blocks of extremal Z2-odd
operators. If everything works, this should be a function that
progresses through crossing equations looking for coefficients in common
with previously solved ones until none are left.

Signed-off-by: Connor Behan <connor.behan@gmail.com>