Warn about bad parameter choices
If the number of derivative components is extreme compared to the number
of poles and spins, crossing symmetry is trivially impossible. Forget
positive OPE coefficients. The blocks will not be able to have the right
sum even with negative ones. This takes the approach of warning the user
after the XML file is generated. An earlier warning may be premature
since the user can add isolated points.

Signed-off-by: Connor Behan <connor.behan@gmail.com>