Add parameter for using overdetermined equations
If squared OPE coefficients are found from an extremal spectrum by
inverting a square matrix, it is very hard to make them positive. It is
better to let the equations outnumber the variables by 1 and perform a
fit with the explicit positivity constraint. This adds a parameter
called nullity for allowing this excess to be 1 or more.

Signed-off-by: Connor Behan <connor.behan@gmail.com>