Make recognition of poles less picky
The conformal blocks that go into the entries of a vectorial sum rule
are supposed to be associated with the same positive prefactor. This
way, the code can just look at one entry and determine the poles for the
whole vector. However, it is nicer if we allow the user to mix and match
tables as long as the ones that carry the wrong positive prefactor will
ultimately be multiplied by zero.

Signed-off-by: Connor Behan <connor.behan@gmail.com>