Specify MPFR constants correctly
If we plug a double having too many digits into a function, the extra
digits will not magically be remembered just because that function is
eval_mpfr(). Really, we have to use RealMPFR() which accepts strings.
This was responsible for table IO that was not reproducible. Under this
bug, a saved table with 1e-100 precision would load back as a table with
1e-16 precision. The XML files obviously had enough digits to play well
with SDPB, but it was equivalent to having k = 20 in the radial
expansion as a hard limit.

Signed-off-by: Connor Behan <connor.behan@gmail.com>