Add a parameter for linking isolated operators
This may be slightly awkward but it is something I should have added
awhile ago. It allows one to force certain operators to always appear
with the same OPE coefficient. When the one having a larger matrix is
added, an optional parameter should be used to refer to a bunch that
have smaller matrices. One ordered pair is matched to another ordered
pair and then we just need to specify a spin / representation, a
dimension and a relative coefficient.

Signed-off-by: Connor Behan <connor.behan@gmail.com>