Primitive support for lists of spins
If we use a single SDP object to write a single XML file that contains
every spin, we can run into memory issues. A workaround is writing one
spin at a time. To do this, we need more than just an integer argument
labelling whether a term in the crossing equation applies to even spins
or odd spins. We need it to be a list stating precisely which spins are
involved. If the user passes such a list, it will now be detected. XML
files that are split by spin are fine for recent versions of SDPB.

Signed-off-by: Connor Behan <connor.behan@gmail.com>