A bit more support for lists of spins
Adding another undocumented feature for experts. If we want to generate
one XML file for each spin (which is the future), it is inefficient to
keep passing ConvolvedBlockTable objects with many spins. Convolving one
spin of a ConformalBlockTable can now be done with an optional spins
argument which is a list.

Signed-off-by: Connor Behan <connor.behan@gmail.com>