Initial support for the table generator on gitlab
The scalar_blocks code is fast so it would be nice to allow these tables
to be imported. So far, the annoying part is having to fit z and zbar
derivatives into a table that was meant for other variables. And also
the strange fact that Sacado::Tay::Taylor fails on 32-bit systems when
the precision is greater than 385.

Signed-off-by: Connor Behan <connor.behan@gmail.com>