Make sure the unit is non-vanishing
Most of the superfluous poles in the imported blocks are fine. But a
pole at delta = 0 is disastrous for our program where we get the
identity block by plugging in delta = 0 rather than starting over from
scratch. Also note that the --order 5 flag shown on the gitlab example
is much too low for real life.

Signed-off-by: Connor Behan <connor.behan@gmail.com>