Give the user a way to use the old uppergamma
Surprisingly, the incomplete gamma function in C is not always faster.
Whether the system is 32-bit or 64-bit appears to make a big difference.
This is likely to improve when MPFR developers update the asymptotic
series in their implementation. However, for now it seems best to
include a commented code block that allows the user to switch back and
forth.

Signed-off-by: Connor Behan <connor.behan@gmail.com>