Explicitly support OPE scans
To maximize OPE coefficients in the multi-correlator case, it is best to
proceed angle by angle. We now support this with a new argument that
accepts a unit vector (along with an argument to minimize instead of
maximize). It is still possible to use the old approach of getting an
unconstrained maximum by choosing an arbitrary direction and looking at
the smallest eigenvalue. However, these bounds are usually not very
useful.

Signed-off-by: Connor Behan <connor.behan@gmail.com>