Use unisolve for root finding
The extremal functional minimizing has been bad for awhile now. It seems
pointless to bring in a library dependency for such a rarely used
function. Instead, we will switch to the unisolve executable and improve
the logic so that it more closely follows David's script. This gives us
a reliable function for extremal_dimensions() and an obscure niche
function for extremal_coefficients(). Users who want to find the
coefficients reliably as well should probably just use David's script.

Signed-off-by: Connor Behan <connor.behan@gmail.com>