Add a sanitizer for the spectrum files
If we are going to tell people to use David's script, we should make it
easier to read off physical dimensions and OPE coefficients. The SDP
object is ideally suited to this because it stores the gaps and
prefactors that users just have to remember when they are using the
script on its own.

Signed-off-by: Connor Behan <connor.behan@gmail.com>