Adapt to changes in scalar_blocks
Watler has been busy. Apart from changes in how the poles are printed,
there is also a new program which writes JSON files to store the blocks
in partial fraction form. The use case is not yet clear to me because it
leaves everything in terms of the radial co-ordinate which makes
convolution painful.

Signed-off-by: Connor Behan <connor.behan@gmail.com>