Read mixed correlator tables from scalar_blocks properly
Demanding that a block for l = 0 not have a pole at delta = 0 is too
fast. In mixed conformal block tables, delta = 0 is a pole and there is
nothing we can do about it. It is only unmixed blocks where a factor of
delta in the denominator is guaranteed to be paired with a factor of
delta in the numerator.

Signed-off-by: Connor Behan <connor.behan@gmail.com>