Prefer a unit operator computed from scratch
Since the beginning, this program has tried to be clever by using
existing block tables to deduce the unit operator. This causes problems
with scalar_blocks and other tables that use pole shifting to reduce the
degree. The pole shifting procedure is designed to approximate blocks in
the continuum, not at dimension zero which is below the usual unitarity
bound.

Signed-off-by: Connor Behan <connor.behan@gmail.com>