Reduce the number of incomplete gamma evaluations
Many of the (slow) calls to uppergamma() were keeping the second
argument the same and only changing the first one by integer values. All
of these choices turn out to be algebraically related so we can speed
things up by a factor of k_max by caching a basis of uppergamma()
evaluations. For the ones that we keep, MPFR hits the most efficient
code path.

Signed-off-by: Connor Behan <connor.behan@gmail.com>