Add parser for autoboot crossing equations
Leaving users on their own to figure out crossing equations and
tediously input the matrices which encode them is not very nice. For at
least four years, the Mathematica package autoboot has been available
for this purpose. Its output can be copied to a string and it makes
sense to implement a parser for this string which extracts crossing
equations. The user must specify basic parameters which will be used for
whatever conformal blocks need to be generated on the fly. She must also
add isolated operators manually. This has been tested on strings that
come from autoboot but very little has been done to make it fail
gracefully on bad strings.

Signed-off-by: Connor Behan <connor.behan@gmail.com>